+++ config/environment.js
@@ -214,10 +214,6 @@ module.exports = function(environment) {
         },
         featureFlagNames: {
             routes: {
-                'guid-node.forks': 'ember_project_forks_page',
-                'guid-registration.forks': 'ember_project_forks_page',
-                'guid-node.analytics.index': 'ember_project_analytics_page',
-                'guid-registration.analytics.index': 'ember_project_analytics_page',
                 'guid-node.registrations': 'ember_project_registrations_page',
                 'settings.tokens': 'ember_user_settings_tokens_page',
                 'settings.tokens.index': 'ember_user_settings_tokens_page',