+++ lib/osf-components/addon/components/panel/component.ts
@@ -1,0 +1,8 @@
+import { layout } from '@ember-decorators/component';
+import Component from '@ember/component';
+
+import template from './template';
+
+@layout(template)
+export default class Panel extends Component {
+}