+++ mirage/serializers/registration-schema.ts
@@ -1,6 +1,7 @@
+import RegistrationSchema from 'ember-osf-web/models/registration-schema';
 import ApplicationSerializer from './application';
 
-export default class RegistrationSchemaSerializer extends ApplicationSerializer {
+export default class RegistrationSchemaSerializer extends ApplicationSerializer<RegistrationSchema> {
     keyForAttribute(attr: string) {
         if (attr === 'schemaNoConflict') {
             return 'schema';